+++ topology-generator/cmd/generator/generator.go
@@ -1,33 +1,26 @@
 package main
 
 import (
-	"context"
-	"encoding/json"
 	"flag"
 	"log"
-	"net/http"
-	"os"
-	"os/signal"
-	"time"
 
-	"github.com/gorilla/mux"
 	"github.com/leandroberetta/mimik/pkg/api"
-	"github.com/leandroberetta/mimik/pkg/resources"
+	"github.com/leandroberetta/mimik/pkg/controller"
 )
-
-const MODE_SERVER = "s"
 
 var name, istioProxyRequestCPU, istioProxyRequestMemory, mimikRequestCPU,
 	mimikRequestMemory, mimikLimitCPU, mimikLimitMemory, image, version, injection, injectionlabel string
 
 var replicas int
 
-var mode, path string
+var mode string
 
 var namespaces, services, connections, randomconnections int
 
+var generatorConfig api.Generator
+
 func init() {
-	flag.StringVar(&mode, "m", MODE_SERVER, "Running mode: l(local) or s(server)")
+	flag.StringVar(&mode, "m", api.MODE_SERVER, "Running Mode: l(local) or s(server)")
 
 	flag.IntVar(&namespaces, "n", 5, "Number of Namespaces created")
 	flag.IntVar(&services, "s", 5, "Number of Services created")
@@ -47,107 +40,28 @@ func init() {
 	flag.StringVar(&injectionlabel, "injectionlabel", "istio-injection:enabled", "Injection Label")
 	flag.IntVar(&replicas, "replica", 1, "Number of Replicas created")
 
-	path, _ = os.Getwd()
+	api.GlobalConfig = api.NewConfigurations(name, istioProxyRequestCPU, istioProxyRequestMemory, mimikRequestCPU,
+		mimikRequestMemory, mimikLimitCPU, mimikLimitMemory, version, image, injection, replicas, injectionlabel)
+
+	generatorConfig = api.Generator{
+		Namespaces:        namespaces,
+		Services:          services,
+		Connections:       connections,
+		RandomConnections: randomconnections,
+	}
 }
 
 func main() {
 	flag.Parse()
 
-	if mode == MODE_SERVER {
-		log.Println("Running in Server mode")
-		r := mux.NewRouter()
-		r.Path("/generate").HandlerFunc(generateTopologyHandler)
-		r.Path("/").Handler(http.FileServer(http.Dir("./ui/build/")))
-		r.PathPrefix("/static/").Handler(http.StripPrefix("/static/", http.FileServer(http.Dir("./ui/build/static"))))
-
-		srv := &http.Server{
-			Addr:    ":8080",
-			Handler: r,
+	if mode == api.MODE_SERVER || mode == "" {
+		if err := controller.RunServer(); err != nil {
+			log.Fatalf("Running Server error: %v", err)
 		}
-
-		log.Println("serving at :8080")
-		if err := srv.ListenAndServe(); err != nil {
-			log.Fatalf("Starting Server error: %v", err)
-		}
-
-		c := make(chan os.Signal, 1)
-		signal.Notify(c, os.Interrupt)
-		<-c
-
-		log.Println("shutting down")
-		ctx, cancel := context.WithTimeout(context.Background(), time.Second*15)
-		defer cancel()
-
-		if err := srv.Shutdown(ctx); err != nil {
-			log.Fatalf("Server shutdown error: %v", err)
-		}
-
-		log.Println("shutdown complete")
-		os.Exit(0)
 	} else {
-		log.Println("Running in Local mode")
-
-		generator := api.Generator{
-			Namespaces:        namespaces,
-			Services:          services,
-			Connections:       connections,
-			RandomConnections: randomconnections,
-		}
-
-		log.Printf("Generating config:\n %+v", generator)
-
-		config := api.NewConfigurations(name, istioProxyRequestCPU, istioProxyRequestMemory, mimikRequestCPU,
-			mimikRequestMemory, mimikLimitCPU, mimikLimitMemory, version, image, injection, replicas, injectionlabel)
-
-		log.Printf("Generating deploy config:\n %+v", config)
-
-		topology := resources.GenerateTopology(generator, config)
-
-		topo, err := json.Marshal(topology)
-		if err != nil {
-			log.Fatalf("Can not marshal struct to yaml: %v", err)
-		}
-
-		log.Printf("Writing yaml to %s", path+"/deploy.json\n")
-
-		f, err := os.OpenFile(path+"/deploy.json", os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0755)
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		defer f.Close()
-
-		_, err = f.Write(topo)
-
-		if err != nil {
-			log.Fatal(err)
+		if err := controller.RunCLI(generatorConfig); err != nil {
+			log.Fatalf("Running CLI error: %v", err)
 		}
 	}
 
-}
-
-func generateTopologyHandler(w http.ResponseWriter, r *http.Request) {
-	generator := api.Generator{}
-
-	decoder := json.NewDecoder(r.Body)
-	err := decoder.Decode(&generator)
-	if err != nil {
-		log.Println(err)
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-	defer r.Body.Close()
-
-	config := api.NewConfigurations(name, istioProxyRequestCPU, istioProxyRequestMemory, mimikRequestCPU,
-		mimikRequestMemory, mimikLimitCPU, mimikLimitMemory, version, image, injection, replicas, injectionlabel)
-
-	topology := resources.GenerateTopology(generator, config)
-
-	w.WriteHeader(http.StatusCreated)
-
-	if err := json.NewEncoder(w).Encode(topology); err != nil {
-		log.Println(err)
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
 }
