+++ topology-generator/pkg/controller/server.go
@@ -1,0 +1,71 @@
+package controller
+
+import (
+	"context"
+	"encoding/json"
+	"log"
+	"net/http"
+	"os"
+	"os/signal"
+	"time"
+
+	"github.com/gorilla/mux"
+	"github.com/leandroberetta/mimik/pkg/api"
+	"github.com/leandroberetta/mimik/pkg/resources"
+)
+
+func RunServer() error {
+	log.Println("Running in Server mode")
+	r := mux.NewRouter()
+	r.Path("/generate").HandlerFunc(generateTopologyHandler)
+	r.Path("/").Handler(http.FileServer(http.Dir("./ui/build/")))
+	r.PathPrefix("/static/").Handler(http.StripPrefix("/static/", http.FileServer(http.Dir("./ui/build/static"))))
+
+	srv := &http.Server{
+		Addr:    ":8080",
+		Handler: r,
+	}
+
+	log.Println("serving at :8080")
+	if err := srv.ListenAndServe(); err != nil {
+		return err
+	}
+
+	c := make(chan os.Signal, 1)
+	signal.Notify(c, os.Interrupt)
+	<-c
+
+	log.Println("shutting down")
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second*15)
+	defer cancel()
+
+	if err := srv.Shutdown(ctx); err != nil {
+		return err
+	}
+
+	log.Println("shutdown complete")
+	return nil
+}
+
+func generateTopologyHandler(w http.ResponseWriter, r *http.Request) {
+	generator := api.Generator{}
+
+	decoder := json.NewDecoder(r.Body)
+	err := decoder.Decode(&generator)
+	if err != nil {
+		log.Println(err)
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
+	defer r.Body.Close()
+
+	topology := resources.GenerateTopology(generator, api.GlobalConfig)
+
+	w.WriteHeader(http.StatusCreated)
+
+	if err := json.NewEncoder(w).Encode(topology); err != nil {
+		log.Println(err)
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
+}
